Create README.md
SALES AGENTS
A company has several sales agents, who make orders for the company's customers and distribute them to them
the products. To carry out the activities, the company offers an application through which the agents can order
products. The application provides each agent of the company with a terminal through which:
• the agent views the list of all the products sold by the company, together with the related prices
and the quantities in stock.
• the agent can order a quantity of a product. After any valid order, all agents
logged in the application will see the updated list of stocks (it is possible that the trigger
registering an order from an agent will lead to an informative message "quantity
insufficient in stock").